make hfst-tokenise usable as a lib; include simple string→string fn
